Skip iteration graphs that put contraction before sparse output
